# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API for mobile app <br/> <br/> Run `php artisan l5-swagger:generate` to generate updated API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/switch-type/list": {
            "get": {
                "tags": [
                    "Switch Type"
                ],
                "summary": "List of switch types",
                "operationId": "switchTypeList",
                "responses": {
                    "200": {
                        "description": "List of switchs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SwitchType"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/switch-unit/list": {
            "get": {
                "tags": [
                    "Switch Unit"
                ],
                "summary": "List of switch units",
                "operationId": "switchUnitList",
                "responses": {
                    "200": {
                        "description": "List of switchs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SwitchUnitResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/switch-unit/{switchUnit}": {
            "get": {
                "tags": [
                    "Switch Unit"
                ],
                "summary": "Retrieve the details of a specific switch unit",
                "description": "Returns the details of the specified switch unit.",
                "operationId": "getSwitchUnit",
                "parameters": [
                    {
                        "name": "switchUnit",
                        "in": "path",
                        "description": "Switch unit ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Switch unit retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SwitchUnitResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Switch unit not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/switch-unit/{switchUnit}/switch-type/list": {
            "get": {
                "tags": [
                    "Switch Unit"
                ],
                "summary": "List of switch types for a switch unit",
                "operationId": "switchUnitSwitchTypeList",
                "parameters": [
                    {
                        "name": "switchUnit",
                        "in": "path",
                        "description": "Switch unit ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of switch types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SwitchType"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/switch-unit/{switchUnit}/pond/{pond}/switches/update/status": {
            "patch": {
                "tags": [
                    "Switch Unit"
                ],
                "summary": "Update the status of all switches in a switch unit",
                "operationId": "switchesStatusUpdate",
                "parameters": [
                    {
                        "name": "pond",
                        "in": "path",
                        "description": "Pond ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "switchUnit",
                        "in": "path",
                        "description": "Switch unit ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "switchesStatus": {
                                        "description": "Array of statuses for the switches (0 for off, 1 for on)",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "enum": [
                                                0,
                                                1
                                            ]
                                        },
                                        "example": [
                                            1,
                                            0,
                                            1,
                                            0,
                                            1,
                                            0,
                                            1,
                                            0,
                                            1,
                                            0,
                                            1,
                                            0
                                        ]
                                    },
                                    "status": {
                                        "description": "Status of the switch unit",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Switches status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Switches status updated successfully"
                                        },
                                        "switchUnit": {
                                            "$ref": "#/components/schemas/SwitchUnitResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "SwitchType": {
                "title": "Switch Type",
                "description": "Represents a sensor type.",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the sensor type.",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the sensor type.",
                        "type": "string"
                    },
                    "slug": {
                        "description": "The slug of the sensor type.",
                        "type": "string"
                    },
                    "description": {
                        "description": "A description of the sensor type.",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SwitchUnitResource": {
                "title": "Switch Unit Resource",
                "required": [
                    "id",
                    "name",
                    "slug",
                    "description",
                    "status",
                    "switches"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the resource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "The name of the resource",
                        "type": "string"
                    },
                    "slug": {
                        "description": "The slug of the resource",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the resource",
                        "type": "string"
                    },
                    "status": {
                        "description": "The status of the resource",
                        "type": "string"
                    },
                    "switches": {
                        "description": "Array of switches",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Switch"
                        }
                    }
                },
                "type": "object"
            },
            "Switch": {
                "required": [
                    "number",
                    "switchType",
                    "status"
                ],
                "properties": {
                    "number": {
                        "description": "The number of the switch",
                        "type": "string"
                    },
                    "switchType": {
                        "description": "The type of the switch",
                        "type": "string"
                    },
                    "status": {
                        "description": "The status of the switch (e.g., on/off)",
                        "type": "string"
                    },
                    "comment": {
                        "description": "Additional comment about the switch",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}